+++ backend/canisters/local_user_index/impl/src/lifecycle/init.rs
@@ -22,6 +22,7 @@ fn init(args: Args) {
         args.user_index_canister_id,
         args.group_index_canister_id,
         args.notifications_canister_id,
+        args.proposals_bot_canister_id,
         args.cycles_dispenser_canister_id,
         args.internet_identity_canister_id,
         canister_pool_target_size,