+++ backend/canisters/proposals_bot/api/src/updates/mod.rs
@@ -1,4 +1,5 @@
 pub mod add_governance_canister;
 pub mod appoint_admins;
+pub mod c2c_submit_proposal;
 pub mod import_proposals_group_into_community;
 pub mod remove_governance_canister;