+++ backend/canisters/proposals_bot/c2c_client/src/lib.rs
@@ -1,0 +1,7 @@
+use canister_client::generate_c2c_call;
+use proposals_bot_canister::*;
+
+// Queries
+
+// Updates
+generate_c2c_call!(c2c_submit_proposal);