+++ backend/canisters/user_index/api/src/lifecycle/init.rs
@@ -1,6 +1,6 @@
 use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
-use types::{BuildVersion, CanisterId, CanisterWasm, UserId};
+use types::{BuildVersion, CanisterId, CanisterWasm};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -11,9 +11,9 @@ pub struct Args {
     pub local_user_index_canister_wasm: CanisterWasm,
     pub group_index_canister_id: CanisterId,
     pub notifications_index_canister_id: CanisterId,
+    pub proposals_bot_canister_id: CanisterId,
     pub cycles_dispenser_canister_id: CanisterId,
     pub storage_index_canister_id: CanisterId,
-    pub proposals_bot_user_id: UserId,
     pub internet_identity_canister_id: CanisterId,
     pub wasm_version: BuildVersion,
     pub test_mode: bool,