+++ frontend/openchat-shared/src/domain/chat/chat.ts
@@ -1427,6 +1427,7 @@ export type SendMessageResponse =
     | SendMessageTooLong
     | SendMessageEmpty
     | TransferCannotBeZero
+    | TransferCannotBeToSelf
     | SendMessageRecipientNotFound
     | TransferFailed
     | TransferLimitExceeded
@@ -1481,6 +1482,10 @@ export type TransferLimitExceeded = {
 
 export type TransferCannotBeZero = {
     kind: "transfer_cannot_be_zero";
+};
+
+export type TransferCannotBeToSelf = {
+    kind: "transfer_cannot_be_to_self";
 };
 
 export type SendMessageRecipientBlocked = {