+++ backend/canisters/proposals_bot/impl/src/jobs/mod.rs
@@ -1,7 +1,13 @@
 use crate::RuntimeState;
 
-pub mod update_finished_proposals;
+mod push_proposals;
+mod retrieve_proposals;
+mod update_finished_proposals;
+mod update_proposals;
 
 pub(crate) fn start(state: &RuntimeState) {
+    push_proposals::start_job_if_required(state);
+    retrieve_proposals::start_job();
     update_finished_proposals::start_job_if_required(state);
+    update_proposals::start_job_if_required(state);
 }