+++ backend/canisters/proposals_bot/impl/src/updates/c2c_submit_proposal.rs
@@ -1,14 +1,16 @@
+use crate::timer_job_types::{SubmitProposalJob, TimerJob};
 use crate::{mutate_state, read_state, RuntimeState};
 use candid::Principal;
 use canister_api_macros::update_msgpack;
 use canister_tracing_macros::trace;
 use local_user_index_canister_c2c_client::{lookup_user, LookupUserError};
 use proposals_bot_canister::c2c_submit_proposal::{Response::*, *};
-use proposals_bot_canister::{ProposalToSubmitAction, Treasury};
+use proposals_bot_canister::{ProposalToSubmit, ProposalToSubmitAction, Treasury};
 use sns_governance_canister::types::manage_neuron::Command;
 use sns_governance_canister::types::proposal::Action;
-use sns_governance_canister::types::{Motion, Proposal, Subaccount, TransferSnsTreasuryFunds};
-use types::{CanisterId, SnsNeuronId};
+use sns_governance_canister::types::{manage_neuron_response, Motion, Proposal, Subaccount, TransferSnsTreasuryFunds};
+use types::{CanisterId, SnsNeuronId, UserId};
+use utils::time::SECOND_IN_MS;
 
 #[update_msgpack]
 #[trace]
@@ -22,35 +24,13 @@ async fn c2c_submit_proposal(args: Args) -> Response {
         Err(response) => return response,
     };
 
-    match lookup_user(caller, local_user_index_canister_id).await {
-        Ok(_) => {}
+    let user_id = match lookup_user(caller, local_user_index_canister_id).await {
+        Ok(u) => u.user_id,
         Err(LookupUserError::UserNotFound) => unreachable!(),
         Err(LookupUserError::InternalError(error)) => return InternalError(error),
-    }
-
-    let make_proposal_args = sns_governance_canister::manage_neuron::Args {
-        subaccount: neuron_id.to_vec(),
-        command: Some(Command::MakeProposal(Proposal {
-            title: args.proposal.title,
-            summary: args.proposal.summary,
-            url: args.proposal.url,
-            action: Some(convert_proposal_action(args.proposal.action)),
-        })),
     };
-    if let Err(error) =
-        sns_governance_canister_c2c_client::manage_neuron(args.governance_canister_id, &make_proposal_args).await
-    {
-        mutate_state(|state| {
-            state.data.fire_and_forget_handler.send(
-                args.governance_canister_id,
-                "manage_neuron".to_string(),
-                candid::encode_one(&make_proposal_args).unwrap(),
-            )
-        });
-        Retrying(format!("{error:?}"))
-    } else {
-        Success
-    }
+
+    submit_proposal(user_id, args.governance_canister_id, neuron_id, args.proposal).await
 }
 
 struct PrepareResult {
@@ -72,6 +52,60 @@ fn prepare(args: &Args, state: &RuntimeState) -> Result<PrepareResult, Response>
         })
     } else {
         Err(GovernanceCanisterNotSupported)
+    }
+}
+
+pub(crate) async fn submit_proposal(
+    user_id: UserId,
+    governance_canister_id: CanisterId,
+    neuron_id: SnsNeuronId,
+    proposal: ProposalToSubmit,
+) -> Response {
+    let make_proposal_args = sns_governance_canister::manage_neuron::Args {
+        subaccount: neuron_id.to_vec(),
+        command: Some(Command::MakeProposal(Proposal {
+            title: proposal.title.clone(),
+            summary: proposal.summary.clone(),
+            url: proposal.url.clone(),
+            action: Some(convert_proposal_action(proposal.action.clone())),
+        })),
+    };
+    match sns_governance_canister_c2c_client::manage_neuron(governance_canister_id, &make_proposal_args).await {
+        Ok(response) => {
+            if let Some(command) = response.command {
+                return match command {
+                    manage_neuron_response::Command::MakeProposal(p) => {
+                        mutate_state(|state| {
+                            state.data.nervous_systems.record_user_submitted_proposal(
+                                governance_canister_id,
+                                user_id,
+                                p.proposal_id.unwrap().id,
+                            )
+                        });
+                        Success
+                    }
+                    manage_neuron_response::Command::Error(error) => InternalError(format!("{error:?}")),
+                    _ => unreachable!(),
+                };
+            }
+            InternalError("Response command was empty".to_string())
+        }
+        Err(error) => {
+            mutate_state(|state| {
+                let now = state.env.now();
+                state.data.timer_jobs.enqueue_job(
+                    TimerJob::SubmitProposal(SubmitProposalJob {
+                        user_id,
+                        governance_canister_id,
+                        neuron_id,
+                        proposal,
+                    }),
+                    now + (10 * SECOND_IN_MS),
+                    now,
+                )
+            });
+            Retrying(format!("{error:?}"))
+        }
     }
 }